refactor: Creates Folder for Weapons
